+++ src/main/java/com/library/dto/response/BookResponseDto.java
@@ -23,5 +23,5 @@ public class BookResponseDto {
 
     private String category;
 
-    private BigDecimal dailyPrice;
+    private Double dailyPrice;
 }
