+++ src/main/java/com/library/entity/Book.java
@@ -31,6 +31,6 @@ public class Book {
 
     private Boolean deleted = false;
 
-    private BigDecimal dailyPrice;
+    private Double dailyPrice;
 
 }
